7-Zip Version 25.01 (renormalized)
- The code for handling symbolic links has been changed
  to provide greater security when extracting files from archives.
  Command line switch -snld20 can be used to bypass default security
  checks when creating symbolic links.

+++ CPP/7zip/UI/Agent/AgentProxy.cpp
@@ -636,7 +636,7 @@ HRESULT CProxyArc2::Load(const CArc &arc, IProgress *progress)
       file.Name = (const wchar_t *)p;
       file.NameLen = 0;
       if (size >= sizeof(wchar_t))
-        file.NameLen = size / sizeof(wchar_t) - 1;
+        file.NameLen = size / (unsigned)sizeof(wchar_t) - 1;
     }
     else
     #endif
